requirements.txt
flask
flask-smorest
flask-sqlalchemy
flask-jwt-extended
flask-cors
marshmallow
requests
python-dotenv
